Restore shared sprite frame crop helper
