+++ src/shared/components/ui/SpriteGenerationModal.tsx
@@ -7,6 +7,7 @@ import { useState, useRef, useMemo, useCallback, useEffect } from "react";
 import { X, Loader2, Check, ImagePlus, Sparkles, ArrowLeft, Crop, RotateCcw } from "lucide-react";
 import { Modal } from "./Modal";
 import { cn } from "../../lib/utils";
+import { cropSpriteDataUrl, type SpriteFrameAdjustments } from "../../lib/sprite-frame-crop";
 import { useUIStore } from "../../stores/ui.store";
 import { spriteApi } from "../../api/image-generation-api";
 import { ImagePromptReviewModal, type ImagePromptOverride, type ImagePromptReviewItem } from "./ImagePromptReviewModal";
@@ -90,13 +91,6 @@ interface SliceAdjustments {
 }
 
 type NumericSliceAdjustmentKey = Exclude<keyof SliceAdjustments, "colCuts" | "rowCuts">;
-
-interface SpriteFrameAdjustments {
-  top: number;
-  bottom: number;
-  left: number;
-  right: number;
-}
 
 type SpriteFrameAdjustmentKey = keyof SpriteFrameAdjustments;
 
@@ -390,41 +384,6 @@ function percentToPixels(size: number, value: number): number {
 
 function clampPercent(value: number): number {
   return Math.max(0, Math.min(80, Number.isFinite(value) ? value : 0));
-}
-
-function cropSpriteDataUrl(dataUrl: string, frame: SpriteFrameAdjustments): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const image = new Image();
-    image.onload = () => {
-      const width = image.naturalWidth;
-      const height = image.naturalHeight;
-      const cropLeft = percentToPixels(width, frame.left);
-      const cropRight = percentToPixels(width, frame.right);
-      const cropTop = percentToPixels(height, frame.top);
-      const cropBottom = percentToPixels(height, frame.bottom);
-      const outputWidth = width - cropLeft - cropRight;
-      const outputHeight = height - cropTop - cropBottom;
-
-      if (outputWidth <= 0 || outputHeight <= 0) {
-        reject(new Error("Frame settings leave no usable sprite area"));
-        return;
-      }
-
-      const canvas = document.createElement("canvas");
-      canvas.width = outputWidth;
-      canvas.height = outputHeight;
-      const ctx = canvas.getContext("2d");
-      if (!ctx) {
-        reject(new Error("Canvas is unavailable"));
-        return;
-      }
-
-      ctx.drawImage(image, cropLeft, cropTop, outputWidth, outputHeight, 0, 0, outputWidth, outputHeight);
-      resolve(canvas.toDataURL("image/png"));
-    };
-    image.onerror = () => reject(new Error("Sprite image could not be loaded"));
-    image.src = dataUrl;
-  });
 }
 
 function clampBoundaries(boundaries: number[], minSpan: number): number[] {
